+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 Release Notes
 =============
+## To Be Decided
+* Web App: Support for adding site extensions with "use_extension"
+
 ## 1.3.0-beta2
 * Deployment Scripts: Specifies cleanup on expiration when retention interval is set, and enables cleanup on success only.
 * Deployment Scripts: Support for running the script after other resources are deployed.